Refactor: entirely migrate to hardhat, refactor test files with ts and typechain (#53)
* complete ft token redpacket unit test

* Refactor: entirely migrate to hardhat, refactor test files with ts and typechain

* use env variable to control workflow of deployment scripts

* style: config eslint with @dimensiondev/eslint-plugin

* modify import method for lodash and add more rules in eslint

* use lodash#times & add new rules in eslint

* simplify usage of BigNumber

* add editorconfig, cspell and modify vscode setting config

* refactor testing files

* resolve rebase master conflict

* use valid type in hardhat configure file and fix error reported by new eslint rules

* remove SafeMath and refactor burnToken

* resolve inconsistent spacing

* fix eslint reported error

+++ contracts/test_token.sol
@@ -5,7 +5,7 @@ pragma solidity >0.4.22;
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract TestToken is ERC20 {
-    constructor(uint initialSupply) ERC20("TestToken", "TEST") {
+    constructor(uint256 initialSupply) ERC20("TestToken", "TEST") {
         _mint(msg.sender, initialSupply);
     }
 }
